requirements.txt
h5py~=3.7.0
numpy==1.23.5
pandas==1.5.2
requests~=2.28.1
pyyaml~=6.0
scikit-image==0.19.3
torchvision~=0.11.0